Refactor SQL queries in GraphQL query modules for improved readability
- Simplified SQL query formatting in `returns.rs`, `sales_orders.rs`, `stock_transfers.rs`, `suppliers.rs`, `tasks.rs`, and `warehouses.rs` by removing unnecessary line breaks.
- Ensured consistent formatting across all query modules for better maintainability.

Add GraphQL model and mutation generation commands

- Introduced `graphql-model.toml` and `graphql-mutation.toml` for generating Rust GraphQL model and mutation files based on database schema.
- Defined conventions for model and mutation generation, including input object structures and resolver methods.

Enhance Gemini CLI command documentation

- Added comprehensive command documentation in `commands.md` to assist users in utilizing the Gemini CLI effectively.
- Included detailed descriptions, usage examples, and argument handling for various commands.

Implement enums for billing models

- Created `enums.rs` in the billing model directory to define various enums related to billing, such as `DisputeStatusEnum`, `InvoiceStatusEnum`, and `PaymentMethodEnum`.
- Ensured enums are compatible with async-graphql and sqlx for seamless integration in the application.

+++ services/graphql-billing/src/models/surcharges.rs
@@ -3,21 +3,20 @@ use std::sync::Arc;
 use async_graphql::dataloader::Loader;
 use chrono::{DateTime, NaiveDate, Utc};
 use graphql_core::PostgresDataLoader;
-use rust_decimal::Decimal;
 use uuid::Uuid;
 
-use super::sea_orm_active_enums::SurchargeCalculationMethodEnum;
+use super::enums::SurchargeCalculationMethodEnum;
 
 #[derive(Debug, Clone, Copy, PartialEq, Hash, Eq)]
 pub struct PrimaryKey(pub Uuid);
 
-#[derive(Clone, Debug, PartialEq, Eq, async_graphql::SimpleObject, sqlx::FromRow)]
+#[derive(Clone, Debug, PartialEq, async_graphql::SimpleObject, sqlx::FromRow)]
 #[graphql(name = "BillingSurcharges")]
 pub struct Model {
     pub id: Uuid,
     pub name: String,
     pub r#type: String,
-    pub amount: Decimal,
+    pub amount: f64,
     pub calculation_method: SurchargeCalculationMethodEnum,
     pub is_active: Option<bool>,
     pub valid_from: Option<NaiveDate>,
